# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RustWeb-Vue API",
    "description": "RustWeb-Vue 全栈管理系统 API 文档",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/qgc/command": {
      "post": {
        "tags": [
          "qgc"
        ],
        "summary": "发送飞控命令",
        "description": "# HTTP 端点\n`POST /api/qgc/command`\n\n# 请求体\n```json\n{ \"command\": \"arm\", \"altitude\": 10 }\n```\n支持命令：`arm` 解锁 / `disarm` 锁定 / `takeoff` 起飞（`altitude` 米）/\n`land` 降落 / `rtl` 返航 / `start` 开始任务 / `pause` 暂停 / `resume` 继续 /\n`click_to_go` 随点随行（`params`=[lat, lon, alt]）/\n`move` 键盘摇杆速度控制（`params`=[vx, vy, vz] 机体速度 m/s）。\n\n# 说明\n命令经 COMMAND_LONG 发送，飞控以 COMMAND_ACK 回执\n（WebSocket `command_ack` 事件，`result_name` 为结果码名称）；\n`click_to_go` / `move` 走 SET_POSITION_TARGET 消息（无回执）。",
        "operationId": "qgcSendCommand",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QgcCommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "命令已发送",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_bool"
                }
              }
            }
          }
        }
      }
    },
    "/api/qgc/config": {
      "get": {
        "tags": [
          "qgc"
        ],
        "summary": "读取配置文件内容",
        "description": "# HTTP 端点\n`GET /api/qgc/config`\n\n# 说明\n返回 `config-qgc.ini` 文件的原始内容，供前端配置编辑器使用。",
        "operationId": "qgcGetConfig",
        "responses": {
          "200": {
            "description": "配置文件内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ConfigContent"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "qgc"
        ],
        "summary": "保存配置文件内容",
        "description": "# HTTP 端点\n`PUT /api/qgc/config`\n\n# 请求体\n`{ \"content\": \"新的 INI 配置内容\" }`\n\n# 说明\n将新配置写入 `config-qgc.ini` 并校验语法；修改的配置在服务重启后生效。",
        "operationId": "qgcSaveConfig",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigContent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "保存成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SavedResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/qgc/mission": {
      "get": {
        "tags": [
          "qgc"
        ],
        "summary": "获取任务快照",
        "description": "# HTTP 端点\n`GET /api/qgc/mission`\n\n# 说明\n返回任务状态（idle / uploading / downloading / clearing）、航点列表\n与飞控当前执行航点序号。下载完成后 `items` 为飞控当前任务\n（首条 seq=0 为首页，真实飞控要求）。",
        "operationId": "qgcGetMission",
        "responses": {
          "200": {
            "description": "任务快照",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_QgcMission"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "qgc"
        ],
        "summary": "上传任务（航点规划）",
        "description": "# HTTP 端点\n`PUT /api/qgc/mission`\n\n# 请求体\n```json\n{ \"items\": [ { \"seq\": 1, \"command\": 16, \"lat\": 31.2304, \"lon\": 121.4737, \"altitude\": 30 } ] }\n```\n\n# 说明\n上传 1~100 个航点（MAV_CMD_NAV_WAYPOINT），序号按数组顺序自动重排；\n服务端自动补首页（seq=0，当前位置）后走 MISSION_COUNT → MISSION_ITEM_INT\n协议上传，进度经 WebSocket `mission_progress` 事件实时推送，\n完成/失败/超时后状态回到 idle。",
        "operationId": "qgcUploadMission",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QgcMissionUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "上传已开始",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_bool"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "qgc"
        ],
        "summary": "清除飞控任务",
        "description": "# HTTP 端点\n`DELETE /api/qgc/mission`\n\n# 说明\n经 MISSION_CLEAR_ALL 协议清除飞控当前任务，飞控以 MISSION_ACK 回执\n（WebSocket `mission_progress` 事件，result=ok 即清除成功）。",
        "operationId": "qgcClearMission",
        "responses": {
          "200": {
            "description": "清除已开始",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_bool"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/qgc/mode": {
      "post": {
        "tags": [
          "qgc"
        ],
        "summary": "切换飞行模式",
        "description": "# HTTP 端点\n`POST /api/qgc/mode`\n\n# 请求体\n```json\n{ \"mode\": \"guided\" }\n```\n\n# 说明\n支持 ArduPilot Copter 模式名（stabilize / acro / alt_hold / auto / guided /\nloiter / rtl / circle / land / drift / sport / poshold / brake / throw），\n也接受数字模式 ID（如 `\"16\"`）。经 SET_MODE 发送（base_mode 带\nCUSTOM_MODE_ENABLED 位，ArduPilot 协议要求）。",
        "operationId": "qgcSetMode",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QgcModeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "模式切换指令已发送",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_bool"
                }
              }
            }
          }
        }
      }
    },
    "/api/qgc/param": {
      "get": {
        "tags": [
          "qgc"
        ],
        "summary": "读取参数表（模拟器维护的 ArduCopter 精简子集）",
        "operationId": "qgcGetParams",
        "responses": {
          "200": {
            "description": "参数表列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_QgcParamList"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "qgc"
        ],
        "summary": "写入单个参数（直达全局参数表，模拟器与 HTTP 共享同一表）",
        "operationId": "qgcSetParam",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QgcParam"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "参数已更新",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_QgcParam"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/qgc/stream": {
      "get": {
        "tags": [
          "qgc"
        ],
        "summary": "读取当前遥测数据流频率（Hz）",
        "operationId": "qgcGetStream",
        "responses": {
          "200": {
            "description": "当前遥测频率",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_QgcStreamResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "qgc"
        ],
        "summary": "设置遥测数据流频率（Hz，运行时调速，模拟器下一拍生效）",
        "operationId": "qgcSetStream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QgcStreamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "已更新遥测频率",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_QgcStreamResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/qgc/telemetry/csv/{name}": {
      "get": {
        "tags": [
          "qgc"
        ],
        "summary": "下载单个遥测 CSV 文件内容（纯文本）",
        "operationId": "qgcGetCsv",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "CSV 文件名",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV 文本内容",
            "content": {
              "text/plain": {}
            }
          },
          "404": {
            "description": "文件不存在"
          }
        }
      }
    },
...
    "/api/qgc/tiles/{z}/{x}/{y}": {
      "get": {
        "tags": [
          "qgc"
        ],
        "summary": "获取地图瓦片（代理 + 磁盘缓存）",
        "description": "# HTTP 端点\n`GET /api/qgc/tiles/{z}/{x}/{y}?token=<JWT>`\n\n# 说明\n浏览器地图引擎（Cesium 图片加载器）无法携带 Authorization 头，token 经 `?token=`\n查询参数在 handler 内校验（同 `/ws`），因此该端点不挂认证中间件。\n\n命中磁盘缓存（`tiles/{z}/{x}/{y}.png`）直接返回，无网络请求（**离线加载**）；\n未命中则从瓦片源（`config-qgc.ini` `[Tiles] Url`，默认高德路网图）\n下载并落盘（**离线保存**：前端批量请求即可保存离线地图包）。\n`404` 表示缓存未命中且瓦片源不可达。",
        "operationId": "qgcGetTile",
        "parameters": [
          {
            "name": "z",
            "in": "path",
            "description": "缩放级别",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "x",
            "in": "path",
            "description": "瓦片列号（Web Mercator）",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "y",
            "in": "path",
            "description": "瓦片行号（Web Mercator）",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "瓦片图片（PNG）",
            "content": {
              "image/png": {}
            }
          },
          "404": {
            "description": "缓存未命中且瓦片源下载失败"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfigContent": {
        "type": "object",
        "description": "配置文件内容",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          }
        }
      },
...
      "QgcCommandRequest": {
        "type": "object",
        "description": "飞控命令请求体（`POST /api/qgc/command`）\n\n# 支持的命令\n- `arm`：解锁（MAV_CMD_COMPONENT_ARM_DISARM, param1=1）\n- `disarm`：锁定（MAV_CMD_COMPONENT_ARM_DISARM, param1=0）\n- `takeoff`：起飞（MAV_CMD_NAV_TAKEOFF, param7=altitude）\n- `land`：降落（MAV_CMD_NAV_LAND）\n- `rtl`：返航（MAV_CMD_NAV_RETURN_TO_LAUNCH）\n- `start`：开始执行任务（MAV_CMD_MISSION_START）\n- `pause`：暂停任务（MAV_CMD_DO_PAUSE_CONTINUE, param1=0）\n- `resume`：继续任务（MAV_CMD_DO_PAUSE_CONTINUE, param1=1）\n- `click_to_go`：随点随行（SET_POSITION_TARGET_GLOBAL_INT，`params`=[lat, lon, alt]）\n- `move`：键盘/摇杆速度控制（SET_POSITION_TARGET_LOCAL_NED，`params`=[vx, vy, vz] 机体速度 m/s）\n\n命令经 COMMAND_LONG 发送，飞控以 COMMAND_ACK 回执（WebSocket `command_ack` 事件）。",
        "required": [
          "command"
        ],
        "properties": {
          "altitude": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "起飞高度（米，仅 takeoff 使用，缺省 10）"
          },
          "command": {
            "type": "string",
            "description": "命令名：arm / disarm / takeoff / land / rtl / start / pause / resume / click_to_go / move"
          },
          "params": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "number",
              "format": "float"
            },
            "description": "附加参数（click_to_go: [lat, lon, alt]；move: [vx, vy, vz]）"
          }
        }
      },
...
      "QgcMissionUploadRequest": {
        "type": "object",
        "description": "任务上传请求体（`PUT /api/qgc/mission`）\n\n上传 1~100 个航点；序号不要求连续，服务端按数组顺序自动重排。",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QgcMissionItem"
            },
            "description": "航点列表（1~100 个）"
          }
        }
      },
...
    }
  }
}